% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_waves.R
\name{parse_waves}
\alias{parse_waves}
\title{parse_waves}
\usage{
parse_waves(
  dataset,
  samps_col = "samps",
  waves_col = "waves",
  abs_col = "abs",
  wave_pairs = list(c(300, 400), c(542, 644))
)
}
\arguments{
\item{dataset}{the dataset being used. must be run through read_nanodrop()}

\item{samps_col}{Column name for samples (default "samps")}

\item{waves_col}{Column name for wavelengths (default "waves")}

\item{abs_col}{Column name for absorbance (default "abs")}

\item{wave_pairs}{List of wavelength pairs to analyze (default: c(300,400) and c(542,644))}
}
\value{
final_combined
}
\description{
Parse wavelength differences for spectral analysis
}
\examples{
test <- parse_waves(dataset = ds1)
}
