% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nanodrop.R
\name{read_nanodrop}
\alias{read_nanodrop}
\title{read_nanodrop}
\usage{
read_nanodrop(
  filepath,
  max_wv = 850,
  min_wv = 190,
  increments = 0.5,
  junk_rows = 10
)
}
\arguments{
\item{max_wv}{max wavelength read by nanodrop}

\item{min_wv}{min wavelength read by nanodrop}

\item{increments}{wavelength increments set on machine}

\item{junk_rows}{number of rows of metadata in the Nanodrop output}

\item{data}{dataset to use}
}
\description{
read_nanodrop
}
