% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_waves.R
\name{parse_waves}
\alias{parse_waves}
\title{parse_waves}
\usage{
parse_waves(
  dataset,
  wv1 = 223,
  wv2 = 270,
  wv3 = 542,
  wv4 = 644,
  formula_int = 0.03467,
  nm_542_coef = 0.02321,
  diff_coef = 0.00271
)
}
\arguments{
\item{dataset}{the dataset being used. must be run through read_nanodrop()}

\item{wv1}{#First peak of interest}

\item{wv2}{#First valley to compare first peak to}

\item{wv3}{#Second peak of interest}

\item{wv4}{#Second valley to compare second peak to}

\item{formula_int}{y intercept value}

\item{nm_542_coef}{multiplier for if the peak is at 542 or not}

\item{diff_coef}{coefficient for the peak height}
}
\value{
final_long
}
\description{
parse_waves
}
\examples{
test <- parse_waves(dataset = ds)
}
