% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_waves.R
\name{parse_waves}
\alias{parse_waves}
\title{parse_waves}
\usage{
parse_waves(
  dataset,
  samps_col = "samps",
  waves_col = "waves",
  abs_col = "abs",
  wave_pairs = list(c(270, 223), c(542, 644))
)
}
\arguments{
\item{dataset}{the dataset being used. must be run through read_nanodrop()}

\item{wv1}{#First peak of interest}

\item{wv2}{#First valley to compare first peak to}

\item{wv3}{#Second peak of interest}

\item{wv4}{#Second valley to compare second peak to}

\item{formula_int}{y intercept value}

\item{nm_542_coef}{multiplier for if the peak is at 542 or not}

\item{diff_coef}{coefficient for the peak height}
}
\value{
final_combined
}
\description{
parse_waves
}
\examples{
test <- parse_waves(dataset = ds1)
}
