% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nanodrop.R
\name{read_nanodrop}
\alias{read_nanodrop}
\title{read_nanodrop}
\usage{
read_nanodrop(
  filepath,
  max_wv = "850.0",
  min_wv = "190.0",
  increments = 0.5,
  junk_rows = 10
)
}
\arguments{
\item{filepath}{}

\item{max_wv}{max wavelength read by nanodrop}

\item{min_wv}{min wavelength read by nanodrop}

\item{increments}{wavelength increments set on machine}

\item{junk_rows}{number of rows of metadata in the Nanodrop output}
}
\value{
ds
}
\description{
read_nanodrop
}
\examples{
read_nanodrop("C:/Users/beabo/OneDrive/Documents/NAU/R Package Dev/nanodRop/data/UV-Vis 5_14_2024 5_03_11 PM.tsv")
}
