% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nanodrop.R
\name{read_nanodrop}
\alias{read_nanodrop}
\title{Read and Process Nanodrop Data}
\usage{
read_nanodrop(filepath, max_wv = "850.0", min_wv = "190.0", error = "e")
}
\arguments{
\item{filepath}{Path to the input TSV file.}

\item{max_wv}{Maximum wavelength value as a string. Default is "850.0".}

\item{min_wv}{Minimum wavelength value as a string. Default is "190.0".}

\item{error}{Error character to check in the "samps" column. Default is "e".}
}
\value{
A data.table with processed data.
}
\description{
Read and Process Nanodrop Data
}
